refactor(deps): resolve circular dependencies (B.4)
- Cycle hybridSearch<->dbSearch: extract SearchMode/HybridSearchResult to
  new searchTypes.ts; repoint hybridSearch/dbSearch/ask imports
- Cycle projectIdentity<->setup: both edges are intentional dynamic
  import() (runtime-safe, no init cycle); add .madgerc skipAsyncImports
  so madge reflects the runtime graph + document the lazy-load sites

madge --circular: 0 cycles. tsc + lint + full suite (1264) green.
