+++ src/lib/dbSearch.ts
@@ -12,7 +12,7 @@
 
 import type { GnosysDB, DbMemory } from "./db.js";
 import type { SearchResult, DiscoverResult } from "./search.js";
-import type { HybridSearchResult, SearchMode } from "./hybridSearch.js";
+import type { HybridSearchResult, SearchMode } from "./searchTypes.js";
 
 // ─── Cosine similarity for inline embeddings ────────────────────────────
