+++ src/lib/hybridSearch.ts
@@ -14,25 +14,9 @@ import type { GnosysResolver, LayeredMemory } from "./resolver.js";
 import { GnosysArchive } from "./archive.js";
 import { GnosysDbSearch } from "./dbSearch.js";
 import type { GnosysDB } from "./db.js";
+import type { HybridSearchResult, SearchMode } from "./searchTypes.js";
 
-export type SearchMode = "keyword" | "semantic" | "hybrid";
-
-export interface HybridSearchResult {
-  relativePath: string;
-  title: string;
-  snippet: string;
-  score: number;
-  /** Which method(s) found this result */
-  sources: ("keyword" | "semantic" | "archive")[];
-  /** Full memory content (loaded on demand for ask engine) */
-  content?: string;
-  /** The memory frontmatter content field */
-  fullContent?: string;
-  /** Memory ID (used for dearchiving) */
-  memoryId?: string;
-  /** Whether this result came from the archive */
-  fromArchive?: boolean;
-}
+export type { HybridSearchResult, SearchMode } from "./searchTypes.js";
 
 /** RRF constant k — standard value from Cormack et al. 2009 */
 const RRF_K = 60;
